+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
 * Record: 修复record有默认数据时,filter返回数据和默认数据混淆在一起的问题
 * Session: start方法支持session关闭后重新启动
 * Request: setServer,setQuery,setPost支持数组参数
+* Request: 增加getReferer快捷方法
 
 ## 0.9.20 (2015-06-18)